Display "cancelled" when cancelling a getobj prompt
Reported by tungtn by email. It's tricky to work out when to do
this, because getobj is a query_key prompt and we use those for
other purposes too; however, if we allow a count, that's a sure
sign that we're trying to parse the result in some way, and so
the quitchars are used for cancelling.