Save file optimizations for monsters
There are two basic optimizations here. One is to change the
encoding for "monster does not know the player's location"; this
gives a decent reduction (a little less than 50%) in the diffs for
monster movement on the current level. The other is to fix a
pessimization in the storage for Elbereth effects on migrating
monsters (and other off-level monsters); we don't want to synch
that to the turn counter, as it doesn't change in a turn-by-turn
way (rather, it gets changed while catching up elapsed time).

This also initializes the dlevel field for migrating monsters to
NULL, rather than leaving it uninitialized, which should help to
catch accidental uses of it.