Paranoia about a potential infinite loop in save loading
This happened when an attempt was made to seek to the last byte of
the header of a file (rather than the first byte after it). I'm not
sure why, but just to make sure this doesn't affect other people,
automatically allow for this happening. (Fixing the cause rather
than the symptom would be better, though.)