Add a new option category, "game creation options"
These are a subcategory of birth options that only affect the new
game sequence, and aren't visible (and aren't saved) ingame.

This fixes a bug that's been around for a while where options like
dogname show as blank in the in-game O screen. (We don't have the
information to know what they were, as they aren't saved in the
save file.) A bonus is that the in-game O screen is now shorter
and less intimidating.

This commit doesn't globals-purge dogname and friends, but makes
doing so possible. (Attempting to purge these is how I noticed the
bug in the first place.)