A start on the -beta3 save system
Not all commands are implemented yet (but we don't generate the
unimplemented ones). This is currently comparable in compression
ratio to the -beta2 save system (possibly slightly better).

Warning: you may want to avoid playtesting this commit if you
value the game. As long as I've got everything right, it won't
break save compatibility. Although I've tested all the codepaths
I can think of offhand, it's possible that a mistake in this
commit will generate corrupt save data that won't be detected by
the desync detector (thus making all your gameplay since the last
save backup, or portion of the game saved with a different save
engine, unrecoverable).