Add a diffing desync detector
Now, whenever we write a new diff, we apply that diff to the old
save file and ensure it produces the current save file. This means
that just as the main desync detector catches round trip errors in
the binary save format, the diffing desync detector catches round
trip errors in the diff format.

There's quite some subtlety here. We can't trigger the check before
writing the diff, because that breaks the invariants on the binary
save location (and the game tries to automatically recover from
that via regenerating the diff, i.e. recovery gets stuck in a loop).
However, triggering the check after writing the diff would with the
old code cause a recovery to the diff we just wrote, thus failing to
fix the problem at all (and maybe introducing silent corruption if
the diffs don't have CRCs). The solution used here is to add a
field to program_state that causes recovery to go back to the old
save's diff, rather than the new one we just wrote. (It isn't 100%
perfect, because if the user chooses manual recovery, they'll then
end up loading the corrupted diff, and the save file will *have* to
be recovered manually. But given that all this code should never
run anyway, and the differ has a much smaller error surface than
the binary save or gamestate, it's good enough for now.)