Reverse the list of timers before loading it
This replaces a quadratic algorithm with a linear one. Profiling
showed that this was one of the top two performance bottlenecks in
a pudding-farming game (which is in turn the sort of game most
demanding in performance).

Pudding farmers might deserve what they get playing local, but I
don't want them to hold the game up for everyone else on a server.