Favour coordinate encoding for monster hitpoints
This takes 2 to 3 extra bytes to get started, then saves one byte
per monster-with-no-extra-data-structure so long as no more than 7
consecutive such monsters in save order are uninjured (if more than
that are uninjured, we lose one byte for the entire string of
monsters, thus bringing us back to parity with increment-based
encoding).

It's possible that this is overfitting for games with a lot of
injured monsters. However, the cost in other games (2-3 bytes per
20 turns) is negligible; we should be optimizing for cases where
there are enough injured monsters that the save cost of an injured
monster is noticeable.

Because mdiff_rle is not implemented yet, this optimization leads
to one of the few cases where zlib compression is still useful: a
string of regenerating monsters gives a string of identical bytes
(ASCII space, as it happens), which zlib can then compress down.