Don't access uninitialized memory in color_on_screen code
The problem is that we need to allow checking to see whether a cell
needs redrawing even before doupdate() has reached it. The solution
used here is to do color_on_screen calculations whenever writing to
nout_win (so that the needed information is always up-to-date), and
also in doupdate() if the palette has changed since the last
doupdate().