More uim/movement-related fixes
Mostly covers the various cases of walking into a wall. We veto
moves into a remembered wall unless the moveonly command is used
(when blind, we print a message letting players know that they can
move into the wall's remembered location, because it might have
been destroyed since). This means that walking into a wall again
costs no time, unless you really want to try hard to bump into it.

Also runs uia code for shift-move commands on the correct square
(we were running it on the player's square, not the intended
destination).