Fix redundant "partly eaten" for certain corpses.
A player saw "a partly eaten the Wizard of Yendor's partly eaten
corpse".  I traced this down to the BUFSZ purge of all things, which
replaced sprintf with msgcat_many but failed to remove the first
argument, a purely mechanical error.

(The sprintf function, typical of C library functions, modifies its
first argument and returns void; whereas, msgcat_many has more modern
semantics of returning an actual value, so it doesn't need to have the
variable that's being set passed as first argument.)

Because the existing prefix was getting included when the prefix was
recalculated, any information already there stayed there, and instead
of "the Wizard of Yendor's partly eaten corpse" we get the redundant
description the player saw.