Teach #genesis the adjectives for various status effects.
I needed to test the effects of something on a sleeping monster, but I
discovered that zapping a wand of sleeping at a monster does not cause
it to sleep.  It causes a different, unrelated effect.  Why?  I have
no idea.  But anyway, I decided it would be useful for debugging if I
could just create sleeping monsters directly.  And then while I was at
it I added support for a number of other status effects as well, so
you can now directly create a fleeing monster, a stunned monster, a
confused monster, a fast or slow monster, etc.  The various
adjectives, both the new ones and the existing tame/hostile/peaceful,
can be combined in any order (except you can't combine ones that
contradict each other, e.g., you can't get a monster that is both tame
and hostile or both fast and slow).