Reworked mstrategy.
mstrategy is now a 8bit field with possible values for different kind of monster actions,
like fleeing/etc similar to the upper 8 of the old field. A target is now instead stored in
mon->(sx|sy). There is also mstratprio for determining how much a monster is willing to offer
to fulfill the current strategy, but it is currently unused.
Also did a few tweaks to displacement refreshing, removed u.dx/dy in favour of using &youmonst
(but it is still not used yet for players), and marked the castle chest specifically to prevent
monsters from looting the castle wand since they can loot containers now.