Associate each message with a message channel
This is the NetHack 4 version of MSGTYPE, at least from the engine
point of view (currently the channels don't do anything, that'll be
in a future commit). The idea is that you can selectively colour or
disable channels that contain messages you don't care about, and in
particular, messages that people will want to customize as a group
will tend to be channelized as a group.

And they said I couldn't do changes that reduce message spam on the
Plane of Fire and touch every pline() call in the game...