Implement message channelization in the API and the client
This involved rewriting the message rendering code on the client.
Again. (For people keeping count, this is the fifth time: 3.4.3,
NitroHack, me (first attempt), dstund, me (this attempt).)
dtsund's code was decent at what it did, but had some cases I
disliked (such as messages that both span and share a line,
something that most NetHack variants avoid), and didn't handle
every corner case. It was also broken on P_MESSAGE pauses, and my
attempts to fix this bug made things worse (which is one of the
things that spurred my rerwite, although not playing nice with
channelization was a big reason too).

I can't guarantee that this one handles every corner case either;
however, more cases are handled here than in any previous system,
including several which would have been broken but nobody had
thought to test. (The key observation is that you can't print a
message onscreen until knowing whether the next message is going
to force a --More-- or not; otherwise, you may use up all the space
on a line and only then discover that you need to print a --More--.
This one of the most notably ugly cases in 3.4.3 wintty, which
sometimes has to print the --More-- outside the message area to
compensate.)

The code itself is inspired by dtsund's, in that it does layout
once when a message arrives (actually, it does layout for each
message when the message after it arrives), and never thereafter
changes it. However, it takes this to larger extremes; in particular,
messages keep their coordinates steady at all times, and instead we
move the region in message-coordinate-space that's drawn.

Currently broken:

  * The window being resized horizontally, especially narrower.
    (Vertical resizes are handled correctly.)

  * Control-P (deimplemented rather than broken).

  * pline_once; this has actually been broken for ages (due to
    not respecting global variables discipline; I thought this
    didn't matter in this particular case, so was in less of a
    hurry to fix it, but it does).

  * --More-- forcing is currently done by both the server and the
    client, with the result that they sometimes get forced twice.

  * There's no way to customize message colours yet. This should
    be quite straightforward in terms of rendering, but it needs
    a user interface.

  * Server cancels sometimes clear messages when they shouldn't
    (but sometimes need to clear messages). This was broken
    beforehand, too. However, at least I understand why it's broken
    now (the correct behaviour depends on what the server is going
    to do in the future, which is why it's hard to come up with a
    good solution; the solution probably involves retroactively
    changing whether messages were acknowledged or not, probably
    via treating it lazily).