Reduce code duplication in the menu rendering code
Having two copies of the code is bad enough, but I'm planning to add
a third sort of scrollable window (for the control-P display) and
three copies would be ridiculous.

This also fixes a number of bugs while I'm there (and makes the code
somewhat less daniel_t than it was before).