A start on collapsible menus
Instead of using the items array directly, go via an intermediate
"visible items" array. That way, you can hide items from the menu,
without losing information on the hidden items.

This also adds a framework for the size of a scrollable to change
over time (which would be necessary for a collapsible menu to work
correctly, as otherwise there'd be blank space at the end of it).