Fix off-level monster teleportation
The testsuite discovered that off-level monsters can be teleported
if we try to place a shopkeeper into a shop that's sufficiently full
of garter snakes (or some similar large-group monster); the snake
gets teleported to leave room for the shopkeeper. The teleportation
code wasn't working for off-level monsters; this patch fixes it so
that monster teleportation takes a level argument.