Fix digestion instadeath while polymorphed
The testsuite managed to catch two bugs for the price of one here.
Digestion instadeath (used to) deal damage equal to your human-form
max HP, piercing half physical damage. That might sound like an
instadeath, but it's not necessarily going to work when polymorphed.
This may well be a bug inherited from NetHack 3.4.3 (I haven't
checked yet).

The other bug was in channelization: if we set the "impending death
of the player" flag, it wasn't cleared during rehumanization. The
situation normally couldn't come up (the flag isn't set by most HP
damage sources), but because digestion is both an instadeath and one
that works via HP damage, the flag was just ending up left in the
set state, rightfully causing neutral_turnstate_tasks to complain.
This bug was introduced very recently, and probably also applies to
Vorpal Blade and the Tsurugi of Muramasa (the same fix handles every
case like this, though: just clearing the flag upon rehumanization).