Handle successful prayers with negative luck
I had a hunch that running the testsuite on Friday 13 would be useful,
although it actually seems to be unrelated in this case.

You can't normally get a successful prayer with negative luck. But you
can if your luck changes during the prayer. The reproduction the test
suite found was to kill a monster, getting a Luck penalty in the
process, during the prayer via use of a scroll of stinking cloud.

The relevant codepaths exist in 3.4.3, so it's likely that this bug
was inherited from there.