Fixed a critical bug with monsters applying bullwhips creating irrecoverable saves
The "irrecoverable save" part is due to a NH4 bug -- a very poorly placed impossible() in restmonchn (restore.c:298) later on causing a divide by zero. As a hotfix, score calculation now simply returns 0 in this case.. I am not exactly sure why score calculation is made in first place in a restoration state.