Fixed pet AI and actual eating disagreeing on what is tainted and not.
This was fixed by creating a new function that address corpse rot status.
The intention is to also use this for the "This corpse smells too old ...",
and to allow for Grunthack-style reporting of corpses as being "slightly rotten", etc,
but addressing this now is a non-bugfix that impacts gameplay, and thus is postponed until after Junethack.