Fix a crash upon a midturn save after a "soft cancel" of PICK_LETTER
Reproduction: press + and a spell letter, press space to cancel back
to the spell menu (i.e. a key that dismisses the menu without
cancelling it and without making a selection), then close the window.
NetHack 4 would segfault upon trying to load the resulting midturn
save.

Pushing now as this is affecting a current Junethack game. Thanks to
oh6 for the bug report.