Reworked flow control of beam wands (and expensive cameras).
Now, they all use a commin bhit_at() function to handle drawbridges, doors, monsters and objects.
I also got rid of the bypass system in favour of flagging objects that was to be hit by a zap. This is not supposed to be set at neutral turnstate -- this can happen if a player dies, but that shouldn't be an issue. Aside from that, it shouldn't happen, because the flag isn't saved to the save file since it's not something that should persist.