Fixed a quirk in how fully-ID-obj-on-death was handled.
It is copying the original obj's memory entirely, including oextra, which confuses ox_* routines.
A better fix is to rework this function to have a "copy obj" function (perhaps reusing parts of
splitobj and make that use it too) and work on that copy, freeing it afterwards. But for now,
just fix the bug.