Reworked dothrow.c routines to be able to handle monsters.
This is part of a bigger rework of monster attack routines to drop mhit[mu] handling of ranged attacks in favour of using muse and player routines for it.
As part of these changes, some refactoring was done in dothrow to remove the m_shot global variable.
The old ranged AI was removed completely from mthrowu. There is currently no replacement, but the idea is to consolidate special abilities into muse AI, including
throwing of weapons (similar to how trice eggs and potions are handled). Currently, there is still logic for selection of ranged weapons, etc, in weapon.c,
which I intend to get rid of.

Some other changes were made as part of this rework:
* Removed almost every function from mthrowu. (thitu is pending removal, the rest is subject to move to more appropriate files, to get rid of mthrowu.c fully).
* Monster proficiencies were added for ranged weapons.
* Added helpers for potential hostile targets in line range, polearm range, and melee range, for pathfinding purposes.
* killer_xname now takes a parameter for whether or not to use an article
* Added killer_msg_mon_obj() to allow formatting as "Killed by a monster's object"
* welded/weldmsg now works on monsters
* Made handling of monster proficiencies a bit saner (no more weird aritmetic, just a bitshift)
* Added limit to struct musable from nh_cmd_arg
* Refined object plural handling
* Fixed bugs in mvismon_at
* Monsters can now benefit from altar BUC checking (their AI wont make use of it at present)
* Pet untaming now frees edog when applicable
* Mjollnir's 1%-fail was removed, it was silly
* Removed handling of ranged combat from mhitu/mhitm/mhitq
* Added mrnl wrapper for rnl that "works" on monsters (Currently not properly implemented on them, it just returns rn2 for them)

The code compiles, but with major refactorings like these, there is bound to be bugs...