Started working on a replacement to special abilities.
Basically, everything that was in the Z menu that wasn't a spell.
I dislike this approach, partially becaause it uses up a bunch of
precious spell letters, but also because it feels like the
completely wrong place to house it.
Now, I am working on reworking how monster abilities work. The
plan is to handle #monster abilities and other non-spell Z things
as a seperate menu, kinda like #technique in Slash'EM, and to
have checks and such and processing of it in common functions.
This will both allow me to consolidate the various special
abilities, and allows me to more easily implement monster use of
them.
So far, I have created an enum to hold ability data as well as a
struct with additional info, and an unfinished abil_usable()
helper.