Encyclopedia database (data.base) entries for most monsters.
Formerly, a comment in data.base claimed that the game code did
not support monster entries, but that was lies.  The game code
supports them just fine, all that needed to be done was actually
writing them.  So I did.  Well, most of them.

I have no doubt that many of the entries I wrote could be improved,
but I hope that these descriptions are better than nothing.

I didn't do the quest-specific monsters, for several reasons:
 1. They are already explained by in-game quest text.
 2. Most new players, the main target audience for this feature,
    are unlikely to encounter them anyway.
 3. There are a lot of them, and their descriptions would be much
    harder to write on average, so it would be almost as much
    work as all the rest of the monsters put together.
 4. Nothing that could be said about them would be of any really
    practical benefit to the player.  It would just be more flavor,
    and really, the quest text has that covered.
 5. They were at the end of the list and I ran out of steam.

So I put in a TODO comment for them and let it go at that.