Added dgl simplemail support from Fourk, with minor tweaks.
The mailbox file is defined as MAILBOXENVAR, default is NHMAILBOX.
You can also define an username that will bypass the "server_messages" game option
by defining MAILOVERRIDE to the username that should bypass it.
This is for administration purposes so that server admins can send a notification
that the server is about to go down or similar.