Cancellation as monk special spell.
Restore Ability is terrible. Not only is the spell not useful, but even if it *was* useful, it's not difficult for monks to cast this spell.

Cancellation is a much better fit, because martial artists come into direct physical contact with magical creatures and thus are highly interested in nullifying those monsters' magical abilities.