Experimental: switch to lz4 compression.
Currently uses the variant that sacrifices compressability for speed.
If saves become too big, there's also a high-compression variant.
It actually compresses *slower* than zlib, but retains the fast decompression,
which was the point of the switchover. But experiment with this first.

Is able to load old saves properly, using zlib compression.
The new compression uses #SIZE# to mark compression (rather than $SIZE$)