Fix dormant problem with AI handling if a monster throws another off-level
This will ensure that if a monster causes another monster to depart the
current level, the monster movement handler doesn't end up iterating through
the wrong level.

I'm not entirely sure if this can happen, but I think a monster breaking
a wand of digging can cause this to happen.