Fix issues with $1-3 sacrifices
Sacrificing $1-3 gold makes the altar god upset, which can be useful if it's
not your god (because it will make your god less angry and increase piety for
an altar gift). This is typical of an upset god. However, it didn't reduce
your Luck like all other cases did, meaning you could do it forever. Fix
this.