Fix zombie status line
Thge zombie status line was using the hardcoded value of 40 instead of
the global TERMINAL_ZOMBIE.  Since the critical level for
zombification was raised to 50, it no longer matched.

Also added one more step in the status indicator for Zombie!!! just to
make it even more obvious when you are about to die.