Minor change to game option change behaviour.
Instead of leaving it up to the client to send a proper command when options
change, leave the responsibility of committing option changes to the save
completely to the game engine.

This should also fix issues with replaymode once I get around to making it
aware of option changes. We also ought to do something similar for when
we rely on topten data.