Overhaul Conflict
Conflict no longer affects the AI. Instead, it causes an area of radius 8
centered on the one causing conflict, to be discordant. Any monster within
this area, excluding monsters on the same tile as the one causing conflict
(can be several in case of steeds, engulfers, etc), will be discorded, forcing
them to attack a random monster next to them in melee. This does not cost
movement points, so it doesn't consume any turns. This attack will not credit
the attacker in any way (experience gain, death drops, angering, etc). Also,
should this result in a kill from the player, it doesn't break pacifist.

Monsters do a willpower save of ring-class strength (even if Conflict can come
from other sources). This is the same as in vanilla.

The discord effect happens at the end of a turn, when everyone has moved.
Monsters/players can only do a single conflicted attack per turn, even if
several monsters are causing conflict in an overlapping area.

Possible future additions:
* teach the AI to abuse Conflict on the player
* Players ought to be able to attain willpower (aka "monster magic resistance").